implementing the generate-reference fn
